[CBOR] Implement support for dateTime types
