+++ src/libraries/System.Security.Cryptography.Encoding/tests/Cbor.Tests/CborReaderTests.Integer.cs
@@ -101,49 +101,6 @@ namespace System.Security.Cryptography.Encoding.Tests.Cbor
         }
 
         [Theory]
-        [InlineData(2, 2, "c202")]
-        [InlineData(0, "2013-03-21T20:04:00Z", "c074323031332d30332d32315432303a30343a30305a")]
-        [InlineData(1, 1363896240, "c11a514b67b0")]
-        [InlineData(23, new byte[] { 1, 2, 3, 4 }, "d74401020304")]
-        [InlineData(32, "http://www.example.com", "d82076687474703a2f2f7777772e6578616d706c652e636f6d")]
-        [InlineData(int.MaxValue, 2, "da7fffffff02")]
-        [InlineData(ulong.MaxValue, new object[] { 1, 2 }, "dbffffffffffffffff820102")]
-        public static void ReadTag_SingleValue_HappyPath(ulong expectedTag, object expectedValue, string hexEncoding)
-        {
-            byte[] encoding = hexEncoding.HexToByteArray();
-            var reader = new CborReader(encoding);
-
-            Assert.Equal(CborReaderState.Tag, reader.Peek());
-            CborTag tag = reader.ReadTag();
-            Assert.Equal(expectedTag, (ulong)tag);
-
-            Helpers.VerifyValue(reader, expectedValue);
-            Assert.Equal(CborReaderState.Finished, reader.Peek());
-        }
-
-        [Theory]
-        [InlineData(new ulong[] { 1, 2, 3 }, 2, "c1c2c302")]
-        [InlineData(new ulong[] { 0, 0, 0 }, "2013-03-21T20:04:00Z", "c0c0c074323031332d30332d32315432303a30343a30305a")]
-        [InlineData(new ulong[] { int.MaxValue, ulong.MaxValue }, 1363896240, "da7fffffffdbffffffffffffffff1a514b67b0")]
-        [InlineData(new ulong[] { 23, 24, 100 }, new byte[] { 1, 2, 3, 4 }, "d7d818d8644401020304")]
-        [InlineData(new ulong[] { 32, 1, 1 }, new object[] { 1, "lorem ipsum" }, "d820c1c182016b6c6f72656d20697073756d")]
-        public static void ReadTag_NestedTags_HappyPath(ulong[] expectedTags, object expectedValue, string hexEncoding)
-        {
-            byte[] encoding = hexEncoding.HexToByteArray();
-            var reader = new CborReader(encoding);
-
-            foreach (ulong expectedTag in expectedTags)
-            {
-                Assert.Equal(CborReaderState.Tag, reader.Peek());
-                CborTag tag = reader.ReadTag();
-                Assert.Equal(expectedTag, (ulong)tag);
-            }
-
-            Helpers.VerifyValue(reader, expectedValue);
-            Assert.Equal(CborReaderState.Finished, reader.Peek());
-        }
-
-        [Theory]
         // all possible definite-length encodings for the value 23
         [InlineData("17")]
         [InlineData("1817")]
@@ -200,16 +157,6 @@ namespace System.Security.Cryptography.Encoding.Tests.Cbor
         }
 
         [Theory]
-        [InlineData("c2")]
-        public static void ReadTag_NoSubsequentData_ShouldPeekEndOfData(string hexEncoding)
-        {
-            byte[] data = hexEncoding.HexToByteArray();
-            var reader = new CborReader(data);
-            reader.ReadTag();
-            Assert.Equal(CborReaderState.EndOfData, reader.Peek());
-        }
-
-        [Theory]
         [InlineData("40")] // empty text string
         [InlineData("60")] // empty byte string
         [InlineData("f6")] // null
@@ -224,67 +171,6 @@ namespace System.Security.Cryptography.Encoding.Tests.Cbor
             InvalidOperationException exn = Assert.Throws<InvalidOperationException>(() => reader.ReadInt64());
 
             Assert.Equal("Data item major type mismatch.", exn.Message);
-        }
-
-        [Theory]
-        [InlineData("40")] // empty text string
-        [InlineData("60")] // empty byte string
-        [InlineData("f6")] // null
-        [InlineData("80")] // []
-        [InlineData("a0")] // {}
-        [InlineData("f97e00")] // NaN
-        [InlineData("fb3ff199999999999a")] // 1.1
-        public static void ReadTag_InvalidTypes_ShouldThrowInvalidOperationException(string hexEncoding)
-        {
-            byte[] data = hexEncoding.HexToByteArray();
-            var reader = new CborReader(data);
-            InvalidOperationException exn = Assert.Throws<InvalidOperationException>(() => reader.ReadTag());
-
-            Assert.Equal("Data item major type mismatch.", exn.Message);
-        }
-
-        [Fact]
-        public static void ReadTag_NestedTagWithMissingPayload_ShouldThrowFormatException()
-        {
-            byte[] data = "9fc2ff".HexToByteArray();
-            var reader = new CborReader(data);
-
-            reader.ReadStartArray();
-            reader.ReadTag();
-            Assert.Equal(CborReaderState.FormatError, reader.Peek());
-            Assert.Throws<FormatException>(() => reader.ReadEndArray());
-        }
-
-        [Theory]
-        [InlineData("8201c202")] // definite length array
-        [InlineData("9f01c202ff")] // equivalent indefinite-length array
-        public static void ReadTag_CallingEndReadArrayPrematurely_ShouldThrowInvalidOperationException(string hexEncoding)
-        {
-            // encoding is valid CBOR, so should not throw FormatException
-            byte[] data = hexEncoding.HexToByteArray();
-            var reader = new CborReader(data);
-
-            reader.ReadStartArray();
-            reader.ReadInt64();
-            reader.ReadTag();
-            Assert.Equal(CborReaderState.UnsignedInteger, reader.Peek());
-            Assert.Throws<InvalidOperationException>(() => reader.ReadEndArray());
-        }
-
-        [Theory]
-        [InlineData("a102c202")] // definite length map
-        [InlineData("bf02c202ff")] // equivalent indefinite-length map
-        public static void ReadTag_CallingEndReadMapPrematurely_ShouldThrowInvalidOperationException(string hexEncoding)
-        {
-            // encoding is valid CBOR, so should not throw FormatException
-            byte[] data = hexEncoding.HexToByteArray();
-            var reader = new CborReader(data);
-
-            reader.ReadStartMap();
-            reader.ReadInt64();
-            reader.ReadTag();
-            Assert.Equal(CborReaderState.UnsignedInteger, reader.Peek());
-            Assert.Throws<InvalidOperationException>(() => reader.ReadEndArray());
         }
 
         [Theory]
