+++ src/libraries/System.Security.Cryptography.Encoding/tests/System.Security.Cryptography.Encoding.Tests.csproj
@@ -52,14 +52,17 @@
     <Compile Include="Cbor.Tests\CborReaderTests.Map.cs" />
     <Compile Include="Cbor.Tests\CborReaderTests.SkipValue.cs" />
     <Compile Include="Cbor.Tests\CborReaderTests.Special.cs" />
+    <Compile Include="Cbor.Tests\CborReaderTests.Tag.cs" />
     <Compile Include="Cbor.Tests\CborWriterTests.Helpers.cs" />
     <Compile Include="Cbor.Tests\CborReaderTests.Array.cs" />
     <Compile Include="Cbor.Tests\CborWriterTests.Array.cs" />
     <Compile Include="Cbor.Tests\CborWriterTests.cs" />
     <Compile Include="Cbor.Tests\CborWriterTests.Map.cs" />
     <Compile Include="Cbor.Tests\CborWriterTests.Special.cs" />
+    <Compile Include="Cbor.Tests\CborWriterTests.Tag.cs" />
     <Compile Include="Cbor\CborReader.SkipValue.cs" />
     <Compile Include="Cbor\CborReader.Special.cs" />
+    <Compile Include="Cbor\CborReader.Tag.cs" />
     <Compile Include="Cbor\CborTag.cs" />
     <Compile Include="Cbor\CborInitialByte.cs" />
     <Compile Include="Cbor\CborReader.Array.cs" />
@@ -82,6 +85,7 @@
     <Compile Include="AsnEncodedData.cs" />
     <Compile Include="AsnEncodedDataCollectionTests.cs" />
     <Compile Include="Base64TransformsTests.cs" />
+    <Compile Include="Cbor\CborWriter.Tag.cs" />
     <Compile Include="Oid.cs" />
     <Compile Include="OidCollectionTests.cs" />
     <Compile Include="PemEncodingFindTests.cs" />
